chore: collapse skills subsystem from 9 to 3 files (#322)
